feat: using u64 instead of usize

+++ sleipnir-bank/src/bank.rs
@@ -260,7 +260,7 @@ pub struct Bank {
     pub millis_per_slot: u64,
 
     // The number of block/slot for which generated transactions can stay valid
-    pub max_age: usize,
+    pub max_age: u64,
 
     // -----------------
     // For TransactionProcessingCallback
@@ -455,8 +455,8 @@ impl Bank {
         // Transaction expiration needs to be a fixed amount of time
         // So we compute how many slot it takes for a transaction to expire
         // Depending on how fast each slot is compute
-        let max_age = (DEFAULT_MS_PER_SLOT * MAX_RECENT_BLOCKHASHES as u64
-            / millis_per_slot) as usize;
+        let max_age = DEFAULT_MS_PER_SLOT * MAX_RECENT_BLOCKHASHES as u64
+            / millis_per_slot;
 
         let mut bank = Self {
             rc: BankRc::new(accounts),
@@ -805,7 +805,7 @@ impl Bank {
         blockhash_queue.get_hash_age(blockhash).map(|age| {
             // Since we don't produce blocks ATM, we consider the current slot
             // to be our block height
-            self.block_height() + blockhash_queue.get_max_age() as u64 - age
+            self.block_height() + blockhash_queue.get_max_age() - age
         })
     }
 
@@ -1248,7 +1248,7 @@ impl Bank {
         max_age: u64,
     ) -> bool {
         let blockhash_queue = self.blockhash_queue.read().unwrap();
-        blockhash_queue.is_hash_valid_for_age(hash, max_age as usize)
+        blockhash_queue.is_hash_valid_for_age(hash, max_age)
     }
 
     // -----------------
